done model mto

+++ 07_django/10_django_many_to_one/articles/views.py
@@ -31,7 +31,9 @@ def create(request):
     if request.method == 'POST':
         form = ArticleForm(request.POST)
         if form.is_valid():
-            article = form.save()
+            article = form.save(commit=False)
+            article.user = request.user
+            article.save()
             return redirect('articles:detail', article.pk)
     else:
         form = ArticleForm()
@@ -44,22 +46,25 @@ def create(request):
 @login_required
 def delete(request, artilce_pk):
     article = Article.objects.get(pk=artilce_pk)
-
-    article.delete()
-
+    if request.user == article.user:
+        article.delete()
     return redirect('articles:index')
 
 
 @login_required
 def update(request, article_pk):
     article = Article.objects.get(pk=article_pk)
-    if request.method == 'POST':
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid():
-            form.save()
-            return redirect('articles:detail', article.pk)
+    # 수정을 요청하는 자 vs 게시글의 작성자 비교
+    if request.user == article.user:
+        if request.method == 'POST':
+            form = ArticleForm(request.POST, instance=article)
+            if form.is_valid():
+                form.save()
+                return redirect('articles:detail', article.pk)
+        else:
+            form = ArticleForm(instance=article)
     else:
-        form = ArticleForm(instance=article)
+        return redirect('articles:index')
     context = {
         'article': article,
         'form': form,
@@ -67,6 +72,7 @@ def update(request, article_pk):
     return render(request, 'articles/update.html', context)
 
 
+@login_required
 def comment_create(request, article_pk):
     # 몇 번 게시글인지 조회
     article = Article.objects.get(pk=article_pk)
@@ -77,6 +83,7 @@ def comment_create(request, article_pk):
         # commit을 False로 지정하면 인스턴스는 반환하면서도 DB에 레코드는 작성하지 않도록 함
         comment = comment_form.save(commit=False)
         comment.article = article
+        comment.user = request.user
         comment.save()
         return redirect('articles:detail', article.pk)
     context = {
@@ -86,10 +93,14 @@ def comment_create(request, article_pk):
     return render(request, 'articles/detail.html', context)
 
 
+@login_required
 def comment_delete(request, article_pk, comment_pk):
     # 삭제할 댓글을 조회
     comment = Comment.objects.get(pk=comment_pk)
     # article_pk = comment.article.pk
-    # 댓글 삭제
-    comment.delete()
+    
+    # 댓글 삭제를 요청하는 자 vs 댓글 작성자
+    if request.user == comment.user:
+        # 댓글 삭제
+        comment.delete()
     return redirect('articles:detail', article_pk)
